+++ crates/aipm-pack/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [0.20.6] - 2026-04-13
+
+## [0.20.5] - 2026-04-13
+
+### Bug Fixes
+- Correct aipm-pack module doc comment to list only implemented commands ([#476](https://github.com/TheLarkInn/aipm/pull/476)) (de6ece8)
+
+## [0.20.4] - 2026-04-12
+
+### Documentation
+- Add instruction file patterns to VS Code LSP document selector ([#466](https://github.com/TheLarkInn/aipm/pull/466)) (8497d8d)
+
 ## [0.20.3] - 2026-04-12
 
 ## [0.20.2] - 2026-04-12